Move same tests to parent class
As HashMapDictionaryTest and LuceneDictionaryTest have a lot of tests
in common, those tests are moved to the parent class "FileBasedDictionaryTest
and both child classes will extend this one.
This gets rid of the duplicate code
