+++ src/test/java/org/aksw/sessa/importing/dictionary/implementation/HashMapDictionaryTest.java
@@ -1,10 +1,14 @@
 package org.aksw.sessa.importing.dictionary.implementation;
 
+import static org.hamcrest.CoreMatchers.hasItem;
+import static org.hamcrest.CoreMatchers.not;
+
 import java.io.IOException;
 import java.util.Set;
 import org.aksw.sessa.helper.files.handler.FileHandlerInterface;
+import org.aksw.sessa.helper.files.handler.ReverseTsvFileHandler;
 import org.aksw.sessa.helper.files.handler.TsvFileHandler;
-import org.aksw.sessa.importing.dictionary.DictionaryInterface;
+import org.aksw.sessa.importing.dictionary.FileBasedDictionary;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -12,35 +16,11 @@ import org.junit.Test;
 /**
  * Created by Simon Bordewisch on 21.06.17.
  */
-public class HashMapDictionaryTest {
-
-  private DictionaryInterface dictionary = null;
-  private final String fileName = "src/test/resources/en_surface_forms_small.tsv";
+public class HashMapDictionaryTest extends FileBasedDictionaryTest {
 
   @Before
   public void init() throws IOException{
-    if (dictionary == null) {
-      FileHandlerInterface handler = new TsvFileHandler(fileName);
+    FileHandlerInterface handler = new TsvFileHandler(TEST_FILE1);
       dictionary = new HashMapDictionary(handler);
-    }
-  }
-
-  //TODO: find a way to avoid code duplication (see LuceneDictionaryTest)
-  @Test
-  public void get_TestTwoWords() {
-    String uri = "http://dbpedia.org/resource/Bill_Gates";
-    Set<String> set = dictionary.get("bill gates");
-    Assert.assertTrue(set.contains(uri));
-  }
-
-  @Test
-  public void get_TestOneWord() {
-    String uri = "http://dbpedia.org/ontology/birthPlace";
-    Set<String> set = dictionary.get("birthplace");
-    Assert.assertTrue(set.contains(uri));
-
-    uri = "http://dbpedia.org/ontology/spouse";
-    set = dictionary.get("wife");
-    Assert.assertTrue(set.contains(uri));
   }
 }
